LVRUBYM-300: Made one service, resource
