+++ app/js/common/services/password.service.js
@@ -2,16 +2,19 @@ var passwordResource = require('./../resources/password.resource.js');
 
 module.exports = angular
   .module('password.service', [
-    passwordResource.name
+    passwordResource.name//,
+    //resetPasswordResource.name
   ])
   .factory('passwordService', passwordService);
 
-passwordService.$inject = ['passwordResource', 'messageService', '$state'];
+passwordService.$inject = ['passwordResource', '$mdToast', '$state'];
 
-function passwordService(passwordResource, messageService, $state) {
+function passwordService(passwordResource, $mdToast, $state) {
   var service = {
     forgot: forgot,
-    reset: reset
+    reset: reset,
+    toggleErrorMsg: toggleErrorMsg
+
   };
   return service;
